core/axwave-mission-backend/requirements.txt
# requirements for mission_backend.py (Python 3.14 compatible in this environment)
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
postgrest==2.28.0
realtime==2.28.0
storage3==0.6.1
httpx==0.28.1
httpcore==1.0.9
h2==4.3.0
python-dotenv==1.2.1
requests==2.32.5
numpy==2.4.2
